requirements.txt
langchain-community >=0.3.24
langchain-huggingface >=0.2.0
langchain >=0.3.25
sentence-transformers >=3.2.1
streamlit >= 1.45.1
faiss-cpu >=1.11.0
accelerate >=1.5.0
hf-xet >=1.1.1
pillow  >= 11.2.1
pandas >=2.1.4
google-genai >=1.16.1
python-dotenv >=1.1.0



